l10n_br_ie_search: merge webservice classes
