+++ l10n_br_ie_search/models/__init__.py
@@ -1,5 +1,4 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 from . import sefaz_webservice
-from . import sintegra_webservice
 from . import res_config_settings
